Add theme detection logic to _app to ensure it always runs (& drill props back down to ThemeSwitcher)
