+++ src/pages/_app.tsx
@@ -8,10 +8,42 @@ import * as React from 'react';
 import '@/styles/globals.css';
 
 export default function App({ Component, pageProps }: AppProps) {
+
+  // Theme detector
+  const [theme, setTheme] = useState<"dark" | "light" | "auto">("auto");
+  useEffect(() => {
+    const detectTheme = () => {
+      if (typeof window !== 'undefined') {
+        if (localStorage.theme === 'dark') {
+          document.documentElement.classList.add('dark');
+          setTheme('dark');
+        } else if (!('theme' in localStorage) && window.matchMedia('(prefers-color-scheme: dark)').matches) {
+          document.documentElement.classList.add('dark');
+          setTheme('auto')
+        } else if (!('theme' in localStorage) && window.matchMedia('(prefers-color-scheme: light)').matches) {
+          document.documentElement.classList.remove('dark');
+          setTheme('auto')
+        } else {
+          document.documentElement.classList.remove('dark');
+          setTheme('light');
+        }
+      }
+    }
+    detectTheme()
+
+    // Detect user preference change
+    window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', detectTheme);
+
+    return () => {
+      window.matchMedia('(prefers-color-scheme: dark)').removeEventListener('change', detectTheme);
+    };
+
+  }, [theme]);
+
   return (
     <WalletContextProvider>
       <div>
-        <Navbar />
+        <Navbar theme={theme} setTheme={setTheme} />
         <Toaster position="bottom-right"
           toastOptions={{
             success: {
